+++ rollingFileAppender_test.go
@@ -36,12 +36,12 @@ func TestRollingAppender(t *testing.T) {
 	pathOne := fmt.Sprintf("%s.log", filepath)
 	info, err := os.Stat(pathOne)
 	assert.Nil(t, err, "Stat should be able to find the log file")
-	assert.Equal(t, info.Size(), 1000, "new file should have 1000 bytes, 500 1's and new lines")
+	assert.Equal(t, info.Size(), int64(1000), "new file should have 1000 bytes, 500 1's and new lines")
 
 	pathTwo := fmt.Sprintf("%s.1.log", filepath)
 	info, err = os.Stat(pathTwo)
 	assert.Nil(t, err, "Stat should be able to find the rolled log file")
-	assert.Equal(t, info.Size(), 2048, "rolled file should have 2048 bytes")
+	assert.Equal(t, info.Size(), int64(2048), "rolled file should have 2048 bytes")
 }
 
 func TestRollingAppenderOneFile(t *testing.T) {
@@ -79,7 +79,7 @@ func TestRollingAppenderOneFile(t *testing.T) {
 	pathOne = fmt.Sprintf("%s.log", filepath)
 	info, err := os.Stat(pathOne)
 	assert.Nil(t, err, "Stat should be able to find the log file")
-	assert.Equal(t, info.Size(), 2548*2, "new file should have all the data, since there isn't any rolling")
+	assert.Equal(t, info.Size(), int64(2548*2), "new file should have all the data, since there isn't any rolling")
 }
 
 func TestRollingAppenderNew(t *testing.T) {
@@ -87,7 +87,7 @@ func TestRollingAppenderNew(t *testing.T) {
 	filepath := path.Join(os.TempDir(), "appendtest")
 	app := NewRollingFileAppender(filepath, "log", int64(100), -1)
 
-	assert.Equal(t, app.maxFiles, 1, "max files defaults to 1")
-	assert.Equal(t, app.maxFileSize, 1024, "max filesize defaults to 1024")
+	assert.Equal(t, app.maxFiles, int16(1), "max files defaults to 1")
+	assert.Equal(t, app.maxFileSize, int64(1024), "max filesize defaults to 1024")
 	assert.Equal(t, app.currentFileName(), fmt.Sprintf("%s.%s", filepath, "log"), "current file name is always prefix.suffix")
 }
